+++ .gitignore
@@ -1,1 +1,3 @@
-/build
+build/
+dpcuda/__pycache__
+disp.png
